+++ src/lib.rs
@@ -120,6 +120,8 @@ pub mod delay;
 pub mod gpio;
 #[cfg(feature = "device-selected")]
 pub mod i2c;
+#[cfg(all(feature = "device-selected", feature = "i2s"))]
+pub mod i2s;
 #[cfg(all(
     feature = "usb_fs",
     any(