+++ CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Update the sdio driver to match the changes in the PAC
 - Update README.md with current information
 
+### Added
+
+- Added support for I2S communication using SPI peripherals, and two examples [#265]
+
+[#265]: https://github.com/stm32-rs/stm32f4xx-hal/pull/265
+
 ## [v0.9.0] - 2021-04-04
 
 ### Changed